fulhaxxxxx

+++ Lab2/calendar.h
@@ -196,7 +196,7 @@ namespace lab2 {
     template <typename S>
     Calendar<T>::Calendar(Calendar<S> const& rhs) {
 
-        if (this != &rhs) {
+        if ((size_t) this != (size_t) &rhs) {
             current_date = T(rhs.current_date);
 
             //Copy assigning events
